testing/20130402.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20130402</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="w3c/testing.20130402.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2013-04-02 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Tue Apr 02 00:00:01 2013</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #testing</li>
 <li id="l-3" _a="heycam|away"><a href="#l-3">#</a> [00:10] * heycam|away is now known as heycam</li>
 <li id="l-4"><a href="#l-4">#</a> [00:23] * Quits: tobie (tobie@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-5"><a href="#l-5">#</a> [00:33] * Quits: ArtB (~abarsto@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-6"><a href="#l-6">#</a> [01:47] * Quits: jhammel (~jhammel@public.cloak) (&quot;leaving&quot;)</li>
 <li id="l-7" _a="heycam"><a href="#l-7">#</a> [03:28] * heycam is now known as heycam|away</li>
 <li id="l-8" _a="heycam|away"><a href="#l-8">#</a> [04:26] * heycam|away is now known as heycam</li>
 <li id="l-9"><a href="#l-9">#</a> [06:18] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-10"><a href="#l-10">#</a> [06:46] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-11"><a href="#l-11">#</a> [06:46] * Quits: plinss (~plinss@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-12"><a href="#l-12">#</a> [06:48] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-13"><a href="#l-13">#</a> [06:48] * Joins: plinss (~plinss@public.cloak)</li>
 <li id="l-14"><a href="#l-14">#</a> [06:51] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-15"><a href="#l-15">#</a> [06:52] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-16"><a href="#l-16">#</a> [07:49] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-17"><a href="#l-17">#</a> [07:52] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-18"><a href="#l-18">#</a> [07:57] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-19"><a href="#l-19">#</a> [08:32] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-20"><a href="#l-20">#</a> [08:39] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-21"><a href="#l-21">#</a> [08:39] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-22" _a="heycam"><a href="#l-22">#</a> [09:02] * heycam is now known as heycam|away</li>
 <li id="l-23"><a href="#l-23">#</a> [09:13] * Quits: timeless (~uid4015@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-24"><a href="#l-24">#</a> [09:14] * Joins: timeless (~uid4015@public.cloak)</li>
 <li id="l-25"><a href="#l-25">#</a> [09:14] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-26"><a href="#l-26">#</a> [09:22] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-27"><a href="#l-27">#</a> [09:35] * Joins: tobie (tobie@public.cloak)</li>
 <li id="l-28"><a href="#l-28">#</a> [09:43] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-29"><a href="#l-29">#</a> [09:43] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-30"><a href="#l-30">#</a> [09:43] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-31"><a href="#l-31">#</a> [10:11] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-32"><a href="#l-32">#</a> [11:31] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-33"><a href="#l-33">#</a> [11:35] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-34"><a href="#l-34">#</a> [11:39] * Quits: jgraham (~jgraham@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-35"><a href="#l-35">#</a> [11:45] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-36"><a href="#l-36">#</a> [12:12] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-37"><a href="#l-37">#</a> [12:25] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-38" _a="abarsto"><a href="#l-38">#</a> [12:25] * abarsto is now known as ArtB</li>
 <li id="l-39"><a href="#l-39">#</a> [12:29] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-40"><a href="#l-40">#</a> [13:22] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-41"><a href="#l-41">#</a> [13:32] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-42"><a href="#l-42">#</a> [13:38] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-43"><a href="#l-43">#</a> [14:36] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-44"><a href="#l-44">#</a> [15:06] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
 <li id="l-45"><a href="#l-45">#</a> [15:07] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-46" _a="ArtB"><a href="#l-46">#</a> [15:25] &lt;ArtB&gt; ms2ger - is there a &quot;preferred&quot; way with testharness to abort running a bunch of tests if the feature to test hasn't been implemented? Looking for an example tests that does something like this.</li>
 <li id="l-47" _a="tobie" _r="ArtB"><a href="#l-47">#</a> [15:30] &lt;tobie&gt; ArtB: what's your use case?</li>
 <li id="l-48" _a="ArtB"><a href="#l-48">#</a> [15:31] &lt;ArtB&gt; tobie <a href="https://dvcs.w3.org/hg/pointerevents/file/default/tests/pointerevents-v1/submissions/Nokia/webidl_pointerevent.html">https://dvcs.w3.org/hg/pointerevents/file/default/tests/pointerevents-v1/submissions/Nokia/webidl_pointerevent.html</a></li>
 <li id="l-49" _a="ArtB"><a href="#l-49">#</a> [15:32] &lt;ArtB&gt; before the test is run, I'd like to check if PointerEvent is implemented and if not, don't run the tests but report PointerEvent isn't implement</li>
 <li id="l-50" _a="ArtB"><a href="#l-50">#</a> [15:32] &lt;ArtB&gt; and by report, I mean use testharness to report this condition</li>
 <li id="l-51" _a="ArtB"><a href="#l-51">#</a> [15:33] &lt;ArtB&gt; I recall some tests report something like NOT RUN</li>
 <li id="l-52" _a="ArtB"><a href="#l-52">#</a> [15:33] &lt;ArtB&gt; so I think I want to tie into that</li>
 <li id="l-53" _a="ArtB"><a href="#l-53">#</a> [15:33] &lt;ArtB&gt; but an example would be nice ;)</li>
 <li id="l-54" _a="tobie"><a href="#l-54">#</a> [15:33] &lt;tobie&gt; These tests should be generated by idlparser, btw.</li>
 <li id="l-55" _a="tobie"><a href="#l-55">#</a> [15:33] &lt;tobie&gt; s/idlparser/idlharness/</li>
 <li id="l-56" _a="Ms2ger"><a href="#l-56">#</a> [15:37] &lt;Ms2ger&gt; That would prevent such silly assertions as assert_equals(typeof ev[name], &quot;DOMString&quot;)</li>
 <li id="l-57" _a="darobin"><a href="#l-57">#</a> [15:39] &lt;darobin&gt; hahaha</li>
 <li id="l-58" _a="tobie" _r="ArtB"><a href="#l-58">#</a> [15:40] &lt;tobie&gt; ArtB: not sure what the best option for this is. Not sure this is something you want to impose on the harness either.</li>
 <li id="l-59" _a="Ms2ger"><a href="#l-59">#</a> [15:41] &lt;Ms2ger&gt; If you don't implement PointerEvent, you should fail all tests...</li>
 <li id="l-60" _a="Ms2ger"><a href="#l-60">#</a> [15:41] &lt;Ms2ger&gt; Well</li>
 <li id="l-61" _a="tobie"><a href="#l-61">#</a> [15:41] &lt;tobie&gt; ArtB:e.g. you could imagine a harness skipping all further tests on a page in case one test fails, if you don't want to pollute your results with too much red.</li>
 <li id="l-62" _a="Ms2ger"><a href="#l-62">#</a> [15:42] &lt;Ms2ger&gt; Not in the sense that people should go out of their way to make tests fail, that's stupid</li>
 <li id="l-63" _a="tobie" _r="Ms2ger"><a href="#l-63">#</a> [15:42] &lt;tobie&gt; Ms2ger: well, the other tests should probably error.</li>
 <li id="l-64" _a="tobie"><a href="#l-64">#</a> [15:42] &lt;tobie&gt; if testharness.js has such a status</li>
 <li id="l-65" _a="Ms2ger"><a href="#l-65">#</a> [15:43] &lt;Ms2ger&gt; I really don't see the point</li>
 <li id="l-66" _a="Ms2ger"><a href="#l-66">#</a> [15:43] &lt;Ms2ger&gt; You run all tests</li>
 <li id="l-67" _a="Ms2ger"><a href="#l-67">#</a> [15:43] &lt;Ms2ger&gt; If the browser doesn't implement the feature, it'll fail a lot of them</li>
 <li id="l-68" _a="Ms2ger"><a href="#l-68">#</a> [15:44] &lt;Ms2ger&gt; So what?</li>
 <li id="l-69" _a="tobie" _r="ArtB"><a href="#l-69">#</a> [15:44] &lt;tobie&gt; ArtB: you also have the other case where you want measurements across implementations to be comparable. In which case you need to run all the tests.</li>
 <li id="l-70" _a="tobie" _r="Ms2ger"><a href="#l-70">#</a> [15:45] &lt;tobie&gt; Ms2ger: either way, this is a decision to be made at the harness level, not at the test level.</li>
 <li id="l-71" _a="ArtB" _r="so"><a href="#l-71">#</a> [15:46] &lt;ArtB&gt; so, some type of precondition function -&gt; if … isn't true, the abort and report that NO tests were run</li>
 <li id="l-72" _a="Ms2ger"><a href="#l-72">#</a> [15:46] &lt;Ms2ger&gt; No</li>
 <li id="l-73" _a="ArtB" _r="well"><a href="#l-73">#</a> [15:47] &lt;ArtB&gt; well, (AFAIK) I can't actually run the PointerEvent tests if PE isn't implemented</li>
 <li id="l-74" _a="tobie" _r="well"><a href="#l-74">#</a> [15:48] &lt;tobie&gt; well, you should be able to.</li>
 <li id="l-75" _a="Ms2ger"><a href="#l-75">#</a> [15:48] &lt;Ms2ger&gt; The tests aren't declared by onload, it seems</li>
 <li id="l-76" _a="tobie"><a href="#l-76">#</a> [15:49] &lt;tobie&gt; They should fail/error/timeout or what not, but they should still run.</li>
 <li id="l-77" _a="tobie" _r="ArtB"><a href="#l-77">#</a> [15:50] &lt;tobie&gt; ArtB: also, why don't you just synthesize a pointer event to test the IIDL?</li>
 <li id="l-78" _a="tobie"><a href="#l-78">#</a> [15:50] &lt;tobie&gt; s/IIDL/IDL/</li>
 <li id="l-79" _a="ArtB"><a href="#l-79">#</a> [15:51] &lt;ArtB&gt; too much work ;)</li>
 <li id="l-80" _a="Ms2ger"><a href="#l-80">#</a> [15:51] &lt;Ms2ger&gt; Well</li>
 <li id="l-81" _a="Ms2ger"><a href="#l-81">#</a> [15:51] &lt;Ms2ger&gt; It would give you a useful test</li>
 <li id="l-82" _a="ArtB" _r="anyhow"><a href="#l-82">#</a> [15:51] &lt;ArtB&gt; anyhow, thanks for the info ms2ger and tobie</li>
 <li id="l-83" _a="Ms2ger"><a href="#l-83">#</a> [15:52] &lt;Ms2ger&gt; At this point, the test could just as well not exist, because nobody will run it</li>
 <li id="l-84" _a="tobie" _r="Well"><a href="#l-84">#</a> [15:52] &lt;tobie&gt; Well, it's a manual test.</li>
 <li id="l-85" _a="Ms2ger"><a href="#l-85">#</a> [15:52] &lt;Ms2ger&gt; Exactly</li>
 <li id="l-86"><a href="#l-86">#</a> [15:57] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-87"><a href="#l-87">#</a> [15:58] * Parts: plh (plehegar@public.cloak) (Leaving)</li>
 <li id="l-88"><a href="#l-88">#</a> [16:02] * Joins: KURDILONDON (~1f3534f2@public.cloak)</li>
 <li id="l-89"><a href="#l-89">#</a> [16:02] * Parts: KURDILONDON (~1f3534f2@public.cloak) (KURDILONDON)</li>
 <li id="l-90"><a href="#l-90">#</a> [16:07] * Joins: darobin_ (rberjon@public.cloak)</li>
 <li id="l-91"><a href="#l-91">#</a> [16:07] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-92" _a="darobin_"><a href="#l-92">#</a> [16:08] * darobin_ is now known as darobin</li>
 <li id="l-93"><a href="#l-93">#</a> [16:18] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-94"><a href="#l-94">#</a> [16:27] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-95"><a href="#l-95">#</a> [16:46] * Quits: zcorpan_ (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-96"><a href="#l-96">#</a> [16:50] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-97"><a href="#l-97">#</a> [16:56] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-98"><a href="#l-98">#</a> [17:12] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-99"><a href="#l-99">#</a> [17:42] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-100"><a href="#l-100">#</a> [17:43] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
 <li id="l-101"><a href="#l-101">#</a> [17:43] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-102"><a href="#l-102">#</a> [17:50] * Joins: darobin_ (rberjon@public.cloak)</li>
 <li id="l-103"><a href="#l-103">#</a> [17:50] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-104"><a href="#l-104">#</a> [17:50] * Quits: zcorpan_ (~zcorpan@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-105" _a="darobin_"><a href="#l-105">#</a> [17:52] * darobin_ is now known as darobin</li>
 <li id="l-106"><a href="#l-106">#</a> [18:26] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-107"><a href="#l-107">#</a> [18:56] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-108"><a href="#l-108">#</a> [18:58] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-109"><a href="#l-109">#</a> [19:10] * Joins: jhammel (~jhammel@public.cloak)</li>
 <li id="l-110"><a href="#l-110">#</a> [19:40] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-111"><a href="#l-111">#</a> [19:48] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-112"><a href="#l-112">#</a> [20:39] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;nn&quot;)</li>
 <li id="l-113"><a href="#l-113">#</a> [20:42] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-114"><a href="#l-114">#</a> [20:50] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-115"><a href="#l-115">#</a> [21:12] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-116"><a href="#l-116">#</a> [21:16] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-117"><a href="#l-117">#</a> [21:38] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-118"><a href="#l-118">#</a> [21:48] * Joins: jgraham (~jgraham@public.cloak)</li>
 <li id="l-119"><a href="#l-119">#</a> [21:59] * Quits: ArtB (~abarsto@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-120"><a href="#l-120">#</a> [22:35] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-121"><a href="#l-121">#</a> [22:35] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-122"><a href="#l-122">#</a> [22:39] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 60 seconds)</li>
 <li id="l-123"><a href="#l-123">#</a> [23:15] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-124"><a href="#l-124">#</a> [23:26] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-125"><a href="#l-125">#</a> Session Close: Wed Apr 03 00:00:00 2013</li>
</ol>
<p id="end">The end :)
